solution: save - СОХРАНЯЮСЬ перед введением СВОЕЙ формы ЛОГИНА
